requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==3.20.1
pymongo==4.1.1
python-dateutil==2.8.2
requests==2.28.0
six==1.16.0
pymongo[srv]
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
